Added senaite.core.api.geo
