+++ src/bika/lims/content/bikasetup.py
@@ -58,9 +58,9 @@ from Products.Archetypes.utils import DisplayList
 from Products.Archetypes.utils import IntDisplayList
 from Products.Archetypes.Widget import RichWidget
 from Products.CMFCore.utils import getToolByName
+from senaite.core.api import geo
 from senaite.core.browser.fields.records import RecordsField
 from senaite.core.interfaces import IHideActionsMenu
-from senaite.core.locales import get_countries
 from senaite.core.p3compat import cmp
 from zope.component import getUtility
 from zope.interface import implements
@@ -920,7 +920,7 @@ class BikaSetup(folder.ATFolder):
             return portal_type
 
     def getCountries(self):
-        items = get_countries()
+        items = geo.get_countries()
         items = map(lambda country: (country.alpha_2, country.name), items)
         return items
